improved docu
