+++ examples/aws/README.md
@@ -1,7 +1,7 @@
 #  Deploy a cluster on AWS
 
 ## Overview
-Use this configuration example to deploy an AWS cluster using Gardener.
+Use the configuration example located in [main.tf](main.tf) to deploy an AWS cluster using Gardener.
 
 Currently, values for the following CIDRs are obtained from the similar clusters deployed on Gardener. See the example:
 ```bash
